Move PdgItem from decay.py to particle.py
