+++ pdg/particle.py
@@ -10,6 +10,90 @@ from pdg.data import PdgData
 from pdg.units import HBAR_IN_GEV_S
 
 
+class PdgItem:
+    def __init__(self, api, pdgitem_id, edition=None):
+        self.api = api
+        self.pdgitem_id = pdgitem_id
+        self.cache = {}
+        self.edition = edition
+
+    def __repr__(self):
+        name = self._get_pdgitem()['name']
+        return 'PdgItem("%s")' % name
+
+    def _get_pdgitem(self):
+        if 'pdgitem' not in self.cache:
+            pdgitem_table = self.api.db.tables['pdgitem']
+            query = select(pdgitem_table).where(pdgitem_table.c.id == bindparam('pdgitem_id'))
+            with self.api.engine.connect() as conn:
+                result = conn.execute(query, {'pdgitem_id': self.pdgitem_id}).fetchone()
+                if result is None:
+                    raise PdgNoDataError('No PDGITEM entry for %s' % self.pdgitem_id)
+                self.cache['pdgitem'] = result._mapping
+        return self.cache['pdgitem']
+
+    def _get_targets(self):
+        pdgitem_map_table = self.api.db.tables['pdgitem_map']
+        query = select(pdgitem_map_table).where(pdgitem_map_table.c.pdgitem_id == bindparam('pdgitem_id'))
+        with self.api.engine.connect() as conn:
+            rows = conn.execute(query, {'pdgitem_id': self.pdgitem_id}).fetchall()
+            for row in rows:
+                yield PdgItem(self.api, row.target_id)
+
+    @property
+    def has_particle(self):
+        if 'has_particle' not in self.cache:
+            pdgparticle_table = self.api.db.tables['pdgparticle']
+            query = select(pdgparticle_table).where(pdgparticle_table.c.pdgitem_id == bindparam('pdgitem_id'))
+            with self.api.engine.connect() as conn:
+                result = conn.execute(query, {'pdgitem_id': self.pdgitem_id}).fetchone()
+                if result:
+                    self.cache['pdgparticle'] = result._mapping
+                    self.cache['has_particle'] = True
+                else:
+                    targets = list(self._get_targets())
+                    if len(targets) == 1 and targets[0].has_particle:
+                        self.cache['pdgparticle'] = targets[0].cache['pdgparticle']
+                        self.cache['has_particle'] = True
+                    else:
+                        self.cache['has_particle'] = False
+        return self.cache['has_particle']
+
+    @property
+    def particle(self):
+        if not self.has_particle:
+            if self.has_particles:
+                raise PdgAmbiguousValueError('No unique PDGPARTICLE for PDGITEM %s' % self.pdgitem_id)
+            raise PdgNoDataError('No PDGPARTICLE for PDGITEM %s' % self.pdgitem_id)
+        p = self.cache['pdgparticle']
+        return PdgParticle(self.api, p['pdgid'], edition=self.edition, set_mcid=p['mcid'])
+
+    @property
+    def particles(self):
+        if self.has_particle:
+            yield self.particle
+        else:
+            for target in self._get_targets():
+                for p in target.particles:
+                    yield p
+
+    @property
+    def has_particles(self):
+        return len(list(self.particles)) > 0
+
+    @property
+    def name(self):
+        return self._get_pdgitem()['name']
+
+    @property
+    def name_tex(self):
+        return self._get_pdgitem()['name_tex']
+
+    @property
+    def item_type(self):
+        return self._get_pdgitem()['item_type']
+
+
 class PdgParticle(PdgData):
     """Container class for all information about a given particle.
